machine-sdk/*-mingw32.conf: Disable SDKTAROPTS
Adding the SDKTAROPTS can result in either a failure or an endless loop when
generating the SDK tarball.  This is due to potential recursive symlinks
within the cross compiler directory structure.

It also means files from places /var/log on the build system can be pulled
into the sdk, dependning on the chosen rootfs of the target system this
is built against.

Disabling it is therefore the best opton, when post processing the tarball,
the symlinks can be dealt with accordingly but the current approach is just
dangerous.

Signed-off-by: Mark Hatle <mark.hatle@windriver.com>
Signed-off-by: Richard Purdie <richard.purdie@linuxfoundation.org>